Update grid websocket to send actual changes down to reduce API load
