+++ packages/server/src/api/controllers/row/utils.ts
@@ -154,3 +154,15 @@ export function cleanExportRows(
 
   return cleanRows
 }
+
+export function getTableId(ctx: any) {
+  if (ctx.request.body && ctx.request.body.tableId) {
+    return ctx.request.body.tableId
+  }
+  if (ctx.params && ctx.params.tableId) {
+    return ctx.params.tableId
+  }
+  if (ctx.params && ctx.params.viewName) {
+    return ctx.params.viewName
+  }
+}
